feat: extract reputation check function using new hypernative screener API
